+++ source/tx_forge.cc
@@ -327,7 +327,7 @@ static void power_law_scale(float *buf, double powscale)
     }
 }
 
-void TX_SpectralSynth(unsigned long long seed, float *buf, int width, double fracdim, double powscale)
+void TX_SpectralSynth(uint64_t seed, float *buf, int width, double fracdim, double powscale)
 {
     SYS_ASSERT(width > 0 && (width & 1) == 0);
     SYS_ASSERT(0 < fracdim && fracdim < 4.0);
